requirements.txt
openai
python-dotenv
pandas
streamlit
selenium
beautifulsoup4
